243184 / Добавлены Роли

+++ app/Http/Controllers/ArticlesController.php
@@ -7,6 +7,7 @@ use App\Contracts\Repositories\CreateArticleServiceContract;
 use App\Contracts\Repositories\UpdateArticleServiceContract;
 use App\Http\Requests\ArticleRequest;
 use App\Http\Requests\TagRequest;
+use App\Models\Article;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -28,6 +29,7 @@ class ArticlesController extends Controller
 
     public function create(): View
     {
+        $this->authorize('create', Article::class);
         return view('pages.articles.create');
     }
 
@@ -39,6 +41,7 @@ class ArticlesController extends Controller
 
     public function store(ArticleRequest $articleRequest, TagRequest $tagRequest): RedirectResponse
     {
+        $this->authorize('create', Article::class);
         $this->createArticleService->create(
             $articleRequest,
             $tagRequest->getTags(),
@@ -51,12 +54,14 @@ class ArticlesController extends Controller
 
     public function edit($slug): View
     {
+        $this->authorize('update', Article::class);
         $article = $this->articleRepository->findBySlug($slug);
         return view('pages.articles.edit', compact('article'));
     }
 
     public function update($slug, ArticleRequest $articleRequest, TagRequest $tagRequest): RedirectResponse
     {
+        $this->authorize('update', Article::class);
         $this->updateArticleService->update(
             $this->articleRepository->findBySlug($slug),
             $articleRequest,
@@ -70,6 +75,7 @@ class ArticlesController extends Controller
 
     public function destroy($article): RedirectResponse
     {
+        $this->authorize('delete', Article::class);
         $this->articleRepository->delete($article);
 
         return redirect()->route('articles.index')
